Complete 01-css-methods exercise
Complete 01-css-methods exercise by creating new css file "style.css"
and coding an external stylesheet for the <div> tag. For the internal
method, add <style> tag into the head to code a stylesheet for the
<p> tag. For the inline method, add "style=" attribute to the existing
<button> tag.

This exercise is to practice methods of implementing stylesheets for
html files and how to code a stylesheet.